refactor App Fetch request into helper utils
