+++ src/utils/buildAddressString.js
@@ -1,0 +1,12 @@
+export const buildAddressString = (provider) => {
+  const ADDRESS_1 = provider["provider address 1"];
+  const ADDRESS_2 = provider["provider address 2"];
+  const COUNTY = provider["provider county"];
+  const TOWN = provider["provider town/city"];
+  const POSTCODE = provider["provider postcode"];
+
+  const addressArray = [ADDRESS_1, ADDRESS_2, COUNTY, TOWN, POSTCODE].filter(
+    (parts) => parts !== "Not Available" && parts
+  );
+  return addressArray.join(", ");
+};
