+++ src/utils/buildLocations.js
@@ -1,0 +1,8 @@
+export const buildLocationSet = (providerData) => {
+  const locationSet = new Set();
+  providerData.forEach((provider) => {
+    console.log(provider);
+    locationSet.add(provider["provider town/city"]);
+  });
+  return locationSet;
+}
